backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pg8000==1.30.3
python-dotenv==1.0.0
